Deduplicate get_lcc_mask
